var, let, delete
the var statement has replaced the def statement.
The let operator can be used to declare variables inside expressions.
The first part of a for loop is now an expression rather than an implicit def statement, allowing you to use the let operator.
The delete keyword is now an operator rather than a statement and returns the value of the deleted variable.
Also fixed some typos in the readme