Iterators are functional
The execution of a statement now returns an enumerable of results.
Most of the time the enumerable will only contain 1 element, but it enables the usage of the yield statement.
If a statement contains yield statements, it can return multiple results.
Also, if a statement does not return a result, it can now return an empty enumerable instead of returning null.