improved try catch statements
catch clause now require parentheses with an identifier to store the caugth exception.
You can optionally add a when clause with an expression which needs to evaluate to true for the exception to be caugth.
You can have multiple catch clause, but you cannot add another one after a catch without a when.